+++ src/components/Images/ImageGallery/ImageGallery.jsx
@@ -1,9 +1,9 @@
 import React from "react";
 import ImageGallery from "react-image-gallery";
 import "react-image-gallery/styles/css/image-gallery.css";
-import Images from './Images';
 
-const Gallery = () => {
+const Gallery = (props) => {
+  const { Images } = props;
   return <ImageGallery
     items={Images}
     autoPlay={true}
