--- src/components/Reviews/Reviews.jsx
@@ -1,9 +1,0 @@
-import React from "react";
-
-export default function Reviews() {
-  return (
-    <>
-      <h1>Reviews</h1>      
-    </>
-  );
-}
