+++ src/features/Email/styles/styles.css
@@ -1,0 +1,24 @@
+.react-datepicker-wrapper {
+  width: 300px;
+  padding-left: 20px;
+}
+
+input {
+  width: 100%;
+  display: block;
+  width: 100%;
+  height: calc(1.5em + 0.75rem + 2px);
+  padding: 0.375rem 0.75rem;
+  font-size: 1rem;
+  font-weight: 400;
+  line-height: 1.5;
+  color: #495057;
+  background-color: #fff;
+  background-clip: padding-box;
+  border: 1px solid #ced4da;
+  border-radius: 0.25rem;
+}
+
+input[type="number"] {
+  width: 100%;
+}
