+++ src/features/Packages/Packages.jsx
@@ -1,14 +1,15 @@
 import React from "react";
-import ImageGallery from '../Images/ImageGallery/ImageGallery';
-import ImageSlider from "../Images/ImageSlider/ImageSlider";
-import VideoCarousel from '../Carousel/Video';
+import ImageGallery from '../../components/Images/ImageGallery/ImageGallery';
+import ImageSlider from "../../components/Images/ImageSlider/ImageSlider";
+import VideoCarousel from '../../components/Video/Video';
+import Images from './Images';
 
 export default function Packages() {
   return (
     <>
       <h1>Packages</h1>
       <div style={{ width: '70%', marginTop: '20px' }}>
-        <ImageGallery />
+        <ImageGallery Images={Images} />
       </div>  
       <div style={{ width: '70%', marginTop: '20px', marginBottom: '20px' }}>
         <ImageSlider />
