+++ src/features/Reviews/Images.js
@@ -1,0 +1,19 @@
+const Images = [
+  {
+    original: "/images/shoe1.jpg",
+    thumbnail: "/images/shoe1.jpg",
+    description : 'Lorem Ipsum is simply dummy text of the printing and typesetting industry.',
+  },
+  {
+    original: "/images/shoe2.jpg",
+    thumbnail: "/images/shoe2.jpg",
+    description : 'Description goes here...',
+  },
+  {
+    original: "/images/shoe3.jpg",
+    thumbnail: "/images/shoe3.jpg",
+    description : 'Description goes here...',
+  },
+];
+
+export default Images;
